refactor: legacy weather and enhanced weather work parallelly depending on toggle
- Add the toggle in opitons->video->misc for enhanced weather
- Add the toggle in opitons->video->misc->enhanced weather for particle effects
- Add a WeatherBase class to support both legace weather and enhanced weather
