+++ src/ClassicUO.Client/Configuration/Profile.cs
@@ -363,7 +363,8 @@ namespace ClassicUO.Configuration
         public int AuraUnderFeetType { get; set => SetProperty(ref field, value); } // 0 = NO, 1 = in warmode, 2 = ctrl+shift, 3 = always
         public bool AuraOnMouse { get; set => SetProperty(ref field, value); } = true;
         public bool AnimatedWaterEffect { get; set => SetProperty(ref field, value); } = false;
-        public bool EnableWeatherEffects { get; set => SetProperty(ref field, value); } = true;
+        public bool EnableWeatherEffects { get; set => SetProperty(ref field, value); } = false;
+        public bool EnableEnhancedWeather { get; set => SetProperty(ref field, value); } = false;
 
         public bool PartyAura { get; set => SetProperty(ref field, value); }
